REFACTOR : Modification du noms des variable
