+++ src/commands/launch/launchEmbedFormater.ts
